[feat][fix] Implement multiple PtlPut/PtlGet operations for scatter-gather lists
This commit modifies RDMA operations to use multiple PtlPut operations for
RDMA writes and multiple PtlGet operations for RDMA reads from Target to
Initiator when a scatter-gather list is defined in the operation.

This approach bypasses the current iovec feature of Portals, which will be
developed and integrated in future board implementations.